Agrego que no se pueda eliminar ni editar una tarea finalizada y que no se pueda editar un proyecto finalizado

+++ src/main/java/com/aninfo/service/TareaService.java
@@ -1,7 +1,10 @@
 package com.aninfo.service;
 
+import com.aninfo.exceptions.ProyectoFinalizadoException;
+import com.aninfo.exceptions.TareaFinalizadaException;
 import com.aninfo.model.Proyecto;
 import com.aninfo.model.Tarea;
+import com.aninfo.repository.ProyectoRepository;
 import com.aninfo.repository.TareaRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
@@ -16,16 +19,33 @@ public class TareaService {
     @Autowired
     private TareaRepository tareaRepository;
 
+    private ProyectoRepository proyectoRepository;
 
     public Collection<Tarea> getTareas(long codigoProyecto) {
         return tareaRepository.getTareasByCodigoProyecto(codigoProyecto);
     }
 
     public Tarea crearTarea(Tarea tarea) {
+        //Proyecto proyecto = proyectoRepository.findProyectoByCodigo(tarea.getCodigoProyecto());
+
+        /*if(proyecto.getEstado().equals("FINALIZADO")){
+            throw new ProyectoFinalizadoException("No se puede agregar una tarea a un proyecto finalizado");
+        }*/
+
         return tareaRepository.save(tarea);
     }
 
     public void deleteById(Long codigoTarea) {
+        Tarea tarea = tareaRepository.findTareaByCodigoTarea(codigoTarea);
+        //Proyecto proyecto = proyectoRepository.findProyectoByCodigo(tarea.getCodigoProyecto());
+
+        /*if(proyecto.getEstado().equals("FINALIZADO")){
+            throw new ProyectoFinalizadoException("No se puede agregar una tarea a un proyecto finalizado");
+        }*/
+
+        if(tarea.getEstado().equals("FINALIZADA")){
+            throw new TareaFinalizadaException("No se puede eliminar una tarea finalizada");
+        }
         tareaRepository.deleteById(codigoTarea);
     }
 
@@ -41,19 +61,29 @@ public class TareaService {
     public void deleteAll() { tareaRepository.deleteAll();
     }
 
-    public boolean existeProyecto(Long codigoTarea) { return tareaRepository.existsById(codigoTarea);
+    public boolean existeTarea(Long codigoTarea) { return tareaRepository.existsById(codigoTarea);
     }
 
     public Tarea modificarTarea(Long codigoTarea, Long codigoProyecto, String nombreTarea, String descripcion, String personaAsignada, String estado) {
         Tarea tarea = tareaRepository.findTareaByCodigoTarea(codigoTarea);
+        //Proyecto proyecto = proyectoRepository.findProyectoByCodigo(tarea.getCodigoProyecto());
+
+        /*if(proyecto.getEstado().equals("FINALIZADO")){
+            throw new ProyectoFinalizadoException("No se puede agregar una tarea a un proyecto finalizado");
+        }*/
+
+        if(tarea.getEstado().equals("FINALIZADA")){
+            throw new TareaFinalizadaException("No se puede modificar una tarea finalizada");
+        }
+
         tarea.setCodigoProyecto(codigoProyecto);
         tarea.setNombre(nombreTarea);
         tarea.setDescripcion(descripcion);
         tarea.setPersonaAsignada(personaAsignada);
         tarea.setEstado(estado);
-        tareaRepository.delete(tarea);
-        tareaRepository.save(tarea);
-        return tarea;
+        tareaRepository.deleteById(tarea.getCodigoTarea());
+
+        return tareaRepository.save(tarea);
     }
 
     public Iterable<Tarea> obtenerTodasLasTareasConConCodigoProyecto(Long codigoProyecto) {
